# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vsy Sports API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /turfs:
    get:
      operationId: listTurfs
      tags: [turfs]
      summary: List all turfs (nearest first)
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of turfs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Turf"
    post:
      operationId: createTurf
      tags: [turfs]
      summary: Create a turf (turf owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTurfBody"
      responses:
        "201":
          description: Turf created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turf"
  /turfs/{id}:
    get:
      operationId: getTurf
      tags: [turfs]
      summary: Get turf details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Turf detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurfDetail"
    put:
      operationId: updateTurf
      tags: [turfs]
      summary: Update turf
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTurfBody"
      responses:
        "200":
          description: Updated turf
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turf"
    delete:
      operationId: deleteTurf
      tags: [turfs]
      summary: Delete turf
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /turfs/{id}/slots:
    get:
      operationId: getTurfSlots
      tags: [turfs]
      summary: Get available slots for a turf on a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
    post:
      operationId: createTurfSlot
      tags: [turfs]
      summary: Create a time slot for a turf
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSlotBody"
      responses:
        "201":
          description: Slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeSlot"
  /turfs/{id}/review:
    post:
      operationId: reviewTurf
      tags: [turfs]
      summary: Leave a review for a turf
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewBody"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # BOOKINGS
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string

    # AUTH
...
    Turf:
      type: object
      required: [id, name, pricePerHour, ownerId]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        pricePerHour:
          type: number
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        area:
          type: string
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        featured:
          type: boolean
        ownerId:
          type: integer
        ownerName:
          type: string
        distanceKm:
          type: number
        createdAt:
          type: string
    TurfDetail:
      allOf:
        - $ref: "#/components/schemas/Turf"
        - type: object
          properties:
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    CreateTurfBody:
      type: object
      required: [name, pricePerHour]
      properties:
        name:
          type: string
        description:
          type: string
        pricePerHour:
          type: number
        images:
          type: array
          items:
            type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        area:
          type: string
        amenities:
          type: array
          items:
            type: string
    TimeSlot:
      type: object
      required: [id, turfId, startTime, endTime, date, isBooked]
      properties:
        id:
          type: integer
        turfId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        date:
          type: string
        isBooked:
          type: boolean
        price:
          type: number
    CreateSlotBody:
      type: object
      required: [startTime, endTime]
      properties:
        startTime:
          type: string
        endTime:
          type: string
        recurringDays:
          type: array
          items:
            type: integer
    Review:
      type: object
      required: [id, turfId, userId, rating]
      properties:
        id:
          type: integer
        turfId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
    ReviewBody:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
        comment:
          type: string

    # BOOKINGS
...